Changing counter to long from int

+++ writer.c
@@ -13,14 +13,14 @@ void* printOutput(void *arg){
     Queue *munch2ToWriter = (Queue *) arg;
     char *string = DequeueString(munch2ToWriter);
 
-    int i = 0;
+    long i = 0;
     while(string != NULL) {
         i++;
         printf("%s\n", string);
         free(string);
         string = DequeueString(munch2ToWriter);
     }
-    printf("TOTAL LINES PROCESSED: %d", i);
+    printf("TOTAL LINES PROCESSED: %ld", i);
 
     pthread_exit(0);
 }
